# docs/api/openapi.yaml
openapi: 3.0.3
info:
  title: Rampart API
  description: |
    Rampart is a lightweight, modern identity and access management server.

    ## API Domains

    - **OAuth 2.0 / OIDC** — Standard protocol endpoints following RFCs
    - **Admin API** — Full resource management (requires admin scope)
    - **Account API** — Self-service for authenticated users
    - **System** — Health and readiness checks

    ## Authentication

    - Admin API: Bearer token with `rampart:admin` scope
    - Account API: Bearer token with valid user session
    - OAuth/OIDC: Per-spec client authentication
  version: 1.0.0
  contact:
    name: Rampart
    url: https://github.com/manimovassagh/rampart
  license:
    name: AGPL-3.0
    url: https://www.gnu.org/licenses/agpl-3.0.html
servers:
  - url: http://localhost:8080
    description: Local development
  - url: https://auth.example.com
    description: Production (example)
...
paths:
...
  /oauth/authorize:
    get:
      tags: [OAuth]
      summary: Authorization endpoint
      description: |
        Initiates an OAuth 2.0 authorization flow. Redirects the user to the
        login page if not already authenticated. Supports Authorization Code
        flow with PKCE (RFC 7636).
      operationId: authorize
      parameters:
        - name: response_type
          in: query
          required: true
          schema:
            type: string
            enum: [code]
        - name: client_id
          in: query
          required: true
          schema:
            type: string
        - name: redirect_uri
          in: query
          required: true
          schema:
            type: string
            format: uri
        - name: scope
          in: query
          required: true
          schema:
            type: string
          description: Space-delimited scopes. Include "openid" for OIDC.
        - name: state
          in: query
          schema:
            type: string
          description: CSRF protection value
        - name: code_challenge
          in: query
          schema:
            type: string
          description: PKCE code challenge (required for public clients)
        - name: code_challenge_method
          in: query
          schema:
            type: string
            enum: [S256]
        - name: nonce
          in: query
          schema:
            type: string
          description: Binds ID token to session (recommended for OIDC)
        - name: prompt
          in: query
          schema:
            type: string
            enum: [none, login, consent, select_account]
        - name: login_hint
          in: query
          schema:
            type: string
      responses:
        "302":
          description: Redirect to login or back to client with authorization code
        "400":
          description: Invalid request (missing or invalid parameters)
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/OAuthError"
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
    clientBasic:
      type: http
      scheme: basic
      description: Client ID and secret via HTTP Basic
    clientPost:
      type: apiKey
      in: query
      name: client_secret
      description: Client secret as form parameter
...
  schemas:
...
    OAuthError:
      type: object
      required: [error]
      properties:
        error:
          type: string
          description: "RFC 6749 error code"
        error_description:
          type: string
...